refactor: Lottos를 LottosManager로 변경
