+++ src/controller/LottoGameController.js
@@ -1,6 +1,6 @@
 import { SETTING, RANKING } from '../constant/setting.js';
 import LottoMachine from '../domain/LottoMachine.js';
-import Lottos from '../domain/Lottos.js';
+import LottosManager from '../domain/LottosManager.js';
 import OutputView from '../view/OutputView.js';
 import InputController from './InputController.js';
 
@@ -14,7 +14,7 @@ class LottoGameController {
 
     const winningNumbers = await InputController.inputWinningNumbers();
     const bonusNumber = await InputController.inputBonusNumber(winningNumbers);
-    this.#lottosWinningResult(winningNumbers, bonusNumber);
+    this.#calculateWinningResult(winningNumbers, bonusNumber);
 
     const restartCommand = await InputController.inputRestartCommand();
     this.#restartGame(restartCommand);
@@ -22,11 +22,11 @@ class LottoGameController {
 
   #createRandomLottos() {
     const lottoList = new LottoMachine(this.#purchaseAmount).getLottoNumberList();
-    this.#lottos = new Lottos(lottoList);
     OutputView.printPurchaseResult(lottoList);
+    this.#lottos = new LottosManager(lottoList);
   }
 
-  #lottosWinningResult(winningNumbers, bonusNumber) {
+  #calculateWinningResult(winningNumbers, bonusNumber) {
     const winningResults = this.#lottos.getWinningResults(winningNumbers, bonusNumber);
     OutputView.printWinningResults(winningResults);
     OutputView.printProfitRate(this.#calculateProfitRate(winningResults));
